Conditionally render things like "delete strat" button if we're creating
If the user just created a new strat, then they'll be brought to the
edit strat page. Since it's brand new, we don't need to render things
like the "delete strat" button on that page. The "cancel" button should
also behave differently: if we're creating a new strat, it should delete
that blank strat which was just created, triggering an action, instead
of just navigating the browser's history back by one.

Perform this conditional rendering based on whether the
"isNewlyCreatedStrat" key/value pair is present on the session cookie.